put common code together
