+++ tls/s2n_record_read.h
@@ -1,0 +1,51 @@
+/*
+ * Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License").
+ * You may not use this file except in compliance with the License.
+ * A copy of the License is located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ * or in the "license" file accompanying this file. This file is distributed
+ * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the License for the specific language governing
+ * permissions and limitations under the License.
+ */
+
+#pragma once
+
+#include "tls/s2n_connection.h"
+
+int s2n_record_parse_aead(const struct s2n_cipher_suite *cipher_suite,
+			  struct s2n_connection *conn,
+			  uint8_t content_type,
+			  uint16_t encrypted_length,
+			  uint8_t *implicit_iv,
+			  struct s2n_hmac_state *mac,
+			  uint8_t *sequence_number,
+			  struct s2n_session_key *session_key);
+int s2n_record_parse_cbc(const struct s2n_cipher_suite *cipher_suite,
+			 struct s2n_connection *conn,
+			 uint8_t content_type,
+			 uint16_t encrypted_length,
+			 uint8_t *implicit_iv,
+			 struct s2n_hmac_state *mac,
+			 uint8_t *sequence_number,
+			 struct s2n_session_key *session_key);
+int s2n_record_parse_composite(const struct s2n_cipher_suite *cipher_suite,
+			       struct s2n_connection *conn,
+			       uint8_t content_type,
+			       uint16_t encrypted_length,
+			       uint8_t *implicit_iv,
+			       struct s2n_hmac_state *mac,
+			       uint8_t *sequence_number,
+			       struct s2n_session_key *session_key);
+int s2n_record_parse_stream(const struct s2n_cipher_suite *cipher_suite,
+			    struct s2n_connection *conn,
+			    uint8_t content_type,
+			    uint16_t encrypted_length,
+			    uint8_t *implicit_iv,
+			    struct s2n_hmac_state *mac,
+			    uint8_t *sequence_number,
+			    struct s2n_session_key *session_key);
